Server now receives request from client and sends the right file according to the clients query. So we now have
client --- "I want the file with name X" ---> server
server --- "Here is file X" ---> client (clients copies reply to disk)

This is basically already not that bad and almost all what
we want. However, we also want to make sure that the sending
of the file from server to client was performed correctly.

So we need to change the second step in above diagram to be

server --- "The checksum of X would be XXXX" ---> client
server --- "Here is file X" ---> client (checks hash, then copies to disk)

Then, the polling of the files from the server is already
quite fault-tolerant!!!

What is still missing in the client then would only be:
- a nice, simple but expressive, fuzzy, query language
  to talk to the server about the right files.
- a way to select a file once the server found more than
  a single file for the clients query. For now, just
  printing "the following files match the request: file1 file2 ..."
  This would then allow the client to reissue the query
  in a more exact fashion. So we don't really need a
  multi-step communication just for this. It is already
  good enough if we can just let the client rerun the request.
- The pushing of files to the server will be handled by another
  process. We could even think about issuing the addition of
  files on another port. Then it is very simple to distinguish
  pushing from polling requests at the server side...