Pushing of files almost done.
Transmission of file from client -> server works, as
the sha1 on the server matches the sha1 on the client.
However, the sha1 is not correctly transmitted from
the client back to the server. Probably, the sendInChunks
method somehow disturbs the read/write communication?

Only need to fix this bug then pushing of files should
work out just fine.