fixed bug by reordering send statements -> uncovered issue in sendInChunks
The issue seems to be the sendInChunks method. The problem is probably
that the overall number of bytes send will almost never be a multiple
of BUFFERSIZE. Therefore, we have the problem that we send
  BUFFERSIZE - (MESSAGESIZE % BUFFERSIZE)
to many bytes (i.e. zero bytes).
These are then fed into other reads later in the program flow. To prevent
this issue, we have to send/receive n * BUFFERSIZE big messages that
we then truncate to back to the actually needed size to avoid these
issues once and for all. This seems to be a general pattern:
always take multiples of the buffer size and then truncate to avoid
these annoying issues of left over message bytes...